+++ app/src/main/AndroidManifest.xml
@@ -13,43 +13,43 @@
         android:theme="@style/Theme.StoreClothes"
         tools:targetApi="31">
         <activity
-            android:name=".ui.WishlistActivity"
+            android:name=".ui.Wishlist.WishlistActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.ProfileActivity"
+            android:name=".ui.Profile.ProfileActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.InformationActivity"
+            android:name=".ui.Profile.InformationActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.ProductDetailActivity"
+            android:name=".ui.Product.ProductDetailActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.OrderActivity"
+            android:name=".ui.Product.OrderActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.RegisterActivity"
+            android:name=".ui.Authentication.RegisterActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.PasswordActivity"
+            android:name=".ui.Authentication.PasswordActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.ForgotActivity"
+            android:name=".ui.Authentication.ForgotActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.SearchActivity"
+            android:name=".ui.Home.SearchActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.HomeActivity"
+            android:name=".ui.Home.HomeActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.FilterByCategoryActivity"
+            android:name=".ui.Home.FilterByCategoryActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.CartActivity"
+            android:name=".ui.Cart.CartActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.LoginActivity"
+            android:name=".ui.Authentication.LoginActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
