+++ app/src/main/java/com/example/storeclothes/ui/Authentication/LoginActivity.java
@@ -1,0 +1,16 @@
+package com.example.storeclothes.ui.Authentication;
+
+import android.os.Bundle;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.example.storeclothes.R;
+
+public class LoginActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_login);
+    }
+}
