+++ app/src/main/java/com/example/storeclothes/ui/Home/HomeActivity.java
@@ -1,4 +1,4 @@
-package com.example.storeclothes.ui;
+package com.example.storeclothes.ui.Home;
 
 import android.os.Bundle;
 
@@ -6,11 +6,11 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import com.example.storeclothes.R;
 
-public class LoginActivity extends AppCompatActivity {
+public class HomeActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_login);
+        setContentView(R.layout.activity_home);
     }
 }
