+++ app/src/main/java/com/example/storeclothes/ui/Home/SearchActivity.java
@@ -1,12 +1,8 @@
-package com.example.storeclothes.ui;
+package com.example.storeclothes.ui.Home;
 
 import android.os.Bundle;
 
-import androidx.activity.EdgeToEdge;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.graphics.Insets;
-import androidx.core.view.ViewCompat;
-import androidx.core.view.WindowInsetsCompat;
 
 import com.example.storeclothes.R;
