+++ app/src/main/res/layout/activity_cart.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.CartActivity">
+    tools:context=".ui.Cart.CartActivity">
     <ScrollView
         android:layout_width="match_parent"
         android:layout_height="0dp"
