+++ app/src/main/res/layout/activity_filter_by_category.xml
@@ -6,7 +6,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.FilterByCategoryActivity">
+    tools:context=".ui.Home.FilterByCategoryActivity">
 
     <ScrollView
         android:layout_width="match_parent"
