+++ app/src/main/res/layout/activity_forgot.xml
@@ -12,7 +12,7 @@
         android:layout_marginStart="16dp"
         android:layout_marginEnd="16dp"
         android:layout_marginTop="24dp"
-        tools:context=".ui.ForgotActivity">
+        tools:context=".ui.Authentication.ForgotActivity">
 
         <com.google.android.material.floatingactionbutton.FloatingActionButton
             android:id="@+id/fabBack"
