+++ app/src/main/res/layout/activity_information.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.InformationActivity">
+    tools:context=".ui.Profile.InformationActivity">
     <ScrollView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
