+++ app/src/main/res/layout/activity_login.xml
@@ -12,7 +12,7 @@
         android:layout_marginStart="16dp"
         android:layout_marginEnd="16dp"
         android:layout_marginTop="24dp"
-        tools:context=".ui.LoginActivity">
+        tools:context=".ui.Authentication.LoginActivity">
 
         <TextView
             android:id="@+id/textViewSignIn"
