+++ app/src/main/res/layout/activity_order.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.OrderActivity">
+    tools:context=".ui.Product.OrderActivity">
 
     <ScrollView
         android:layout_width="match_parent"
