+++ app/src/main/res/layout/activity_password.xml
@@ -12,7 +12,7 @@
         android:layout_marginStart="16dp"
         android:layout_marginEnd="16dp"
         android:layout_marginTop="24dp"
-        tools:context=".ui.PasswordActivity">
+        tools:context=".ui.Authentication.PasswordActivity">
 
         <TextView
             android:id="@+id/textViewSignIn"
