+++ app/src/main/res/layout/activity_product_detail.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.ProductDetailActivity">
+    tools:context=".ui.Product.ProductDetailActivity">
 
     <!-- NestedScrollView chứa nội dung chính -->
     <androidx.core.widget.NestedScrollView
