+++ app/src/main/res/layout/activity_wishlist.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.WishlistActivity">
+    tools:context=".ui.Wishlist.WishlistActivity">
 
     <ScrollView
         android:layout_width="match_parent"
